requirements.txt
apscheduler>=3.10.4
werkzeug>=3.0.0
fastapi
uvicorn
python-multipart
psycopg2-binary
python-dotenv
sqlalchemy